backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
fastembed==0.3.6
onnxruntime==1.18.1
numpy==1.26.4
httpx==0.27.2
pydantic-settings==2.5.2
